cli: introduce support for lima-kilo latest isolation setup
The lima-kilo setup is now using a "root" directory in which everything is
confined.

When using the CLI to create a job, the kubeconfig is read both from "outside"
k8s and from the inside.
 * from the outside so the CLI can auth to the jobs-api
 * from inside so jobs-api can create the job authenticating to k8s on behalf of the user

The kubeconfig hardcodes the path to the TLS certs, an absolute path.

When in the lima-kilo setup, we would need the kubeconfig TLS certs to have different
paths when read from inside/outside, and that's very difficult to achieve.

An easier solution is to just introduce a couple new arguments to override the TLS cert/key
when reading them from outside, which allows to leave the rest of the lima-kilo setup as-is.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I0768f4ea848add6569169feaf241a5748d71cda4